Convert to package. Isolate AnonUser
* Convert collect.py module to collect package to enable isolation of
  classes and easier maintenance of code.
